backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.4
sqlalchemy==2.0.27
aiosqlite==0.19.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
rapidfuzz==3.6.1
pytesseract==0.3.10

qrcode==7.4.2
